fix: add tests and upgrade procedure
