+++ src/cmd/migrate.test.ts
@@ -1,20 +1,7 @@
 import { globSync } from 'glob'
-import {
-  applyChanges,
-  Changes,
-  detectAndRestartOutdatedIstioSidecars,
-  filterChanges,
-  getBuildName,
-  getDeploymentNameFromReplicaSet,
-  policiesMigration,
-  restartCluster,
-  restartDeployment,
-  restartPodOwner,
-  restartStatefulSet,
-} from 'src/cmd/migrate'
+import { applyChanges, Changes, filterChanges, getBuildName, policiesMigration } from 'src/cmd/migrate'
 import { env } from '../common/envalid'
 import { getFileMap } from '../common/repo'
-import { CoreV1Api, V1OwnerReference } from '@kubernetes/client-node'
 import { terminal } from '../common/debug'
 
 // Mock external dependencies at the top level - BEFORE imports
@@ -700,534 +687,6 @@ describe('teamSettingsMigration', () => {
     const expectedValues = getTeamSettingsExpectedValues()
     expect(deps.writeValues).toBeCalledWith(expectedValues, true)
   }, 20000)
-})
-
-describe('getDeploymentNameFromReplicaSet', () => {
-  it('should extract deployment name from valid ReplicaSet name with 8-character hash', () => {
-    const result = getDeploymentNameFromReplicaSet('my-deployment-abcd1234')
-    expect(result).toBe('my-deployment')
-  })
-
-  it('should extract deployment name from valid ReplicaSet name with 10-character hash', () => {
-    const result = getDeploymentNameFromReplicaSet('my-deployment-abcd123456')
-    expect(result).toBe('my-deployment')
-  })
-
-  it('should extract deployment name from complex deployment name', () => {
-    const result = getDeploymentNameFromReplicaSet('team-alpha-service-frontend-abcd1234')
-    expect(result).toBe('team-alpha-service-frontend')
-  })
-
-  it('should return null for invalid ReplicaSet name format', () => {
-    const result = getDeploymentNameFromReplicaSet('invalid-name')
-    expect(result).toBeNull()
-  })
-
-  it('should return null for ReplicaSet name with hash too short', () => {
-    const result = getDeploymentNameFromReplicaSet('my-deployment-abc123')
-    expect(result).toBeNull()
-  })
-
-  it('should return null for ReplicaSet name with hash too long', () => {
-    const result = getDeploymentNameFromReplicaSet('my-deployment-abcd123456789')
-    expect(result).toBeNull()
-  })
-
-  it('should return null for ReplicaSet name with uppercase in hash', () => {
-    const result = getDeploymentNameFromReplicaSet('my-deployment-ABCD1234')
-    expect(result).toBeNull()
-  })
-
-  it('should handle deployment names with hyphens correctly', () => {
-    const result = getDeploymentNameFromReplicaSet('my-complex-deployment-name-abcd1234')
-    expect(result).toBe('my-complex-deployment-name')
-  })
-})
-
-describe('restartStatefulSet', () => {
-  let mockOwnerRef: V1OwnerReference
-  let mockRestartedDeployments: Set<string>
-  let mockD: any
-
-  beforeEach(() => {
-    mockOwnerRef = {
-      kind: 'StatefulSet',
-      name: 'test-statefulset',
-      apiVersion: 'apps/v1',
-      uid: 'test-uid',
-    }
-    mockRestartedDeployments = new Set<string>()
-    mockD = {
-      info: jest.fn(),
-    }
-    jest.clearAllMocks()
-  })
-
-  it('should restart StatefulSet when not in dry run mode', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    const mockExec = jest.fn().mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 })
-    const { $ } = require('zx')
-    ;($ as jest.Mock).mockImplementation(mockExec)
-
-    await restartStatefulSet(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-statefulset')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith('Restarting StatefulSet test-statefulset in namespace test-namespace')
-    expect(mockD.info).toHaveBeenCalledWith('Successfully restarted StatefulSet test-statefulset')
-  })
-
-  it('should not restart StatefulSet in dry run mode', async () => {
-    const mockParsedArgs = { dryRun: true, local: false }
-
-    await restartStatefulSet(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-statefulset')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart StatefulSet test-statefulset in namespace test-namespace',
-    )
-  })
-
-  it('should not restart StatefulSet in local mode', async () => {
-    const mockParsedArgs = { dryRun: false, local: true }
-
-    await restartStatefulSet(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-statefulset')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart StatefulSet test-statefulset in namespace test-namespace',
-    )
-  })
-
-  it('should skip restart if StatefulSet already restarted', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    mockRestartedDeployments.add('test-namespace/test-statefulset')
-
-    await restartStatefulSet(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockD.info).not.toHaveBeenCalled()
-  })
-})
-
-describe('restartCluster', () => {
-  let mockOwnerRef: V1OwnerReference
-  let mockRestartedDeployments: Set<string>
-  let mockD: any
-
-  beforeEach(() => {
-    mockOwnerRef = {
-      kind: 'Cluster',
-      name: 'test-cluster',
-      apiVersion: 'postgresql.cnpg.io/v1',
-      uid: 'test-uid',
-    }
-    mockRestartedDeployments = new Set<string>()
-    mockD = {
-      info: jest.fn(),
-    }
-    jest.clearAllMocks()
-  })
-
-  it('should restart CloudNativePG Cluster when not in dry run mode', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    const mockExec = jest.fn().mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 })
-    const { $ } = require('zx')
-    ;($ as jest.Mock).mockImplementation(mockExec)
-
-    await restartCluster(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-cluster')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith('Restarting CloudNativePG Cluster test-cluster in namespace test-namespace')
-    expect(mockD.info).toHaveBeenCalledWith('Successfully initiated rolling restart for Cluster test-cluster')
-  })
-
-  it('should not restart Cluster in dry run mode', async () => {
-    const mockParsedArgs = { dryRun: true, local: false }
-
-    await restartCluster(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-cluster')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart CloudNativePG Cluster test-cluster in namespace test-namespace',
-    )
-  })
-
-  it('should not restart Cluster in local mode', async () => {
-    const mockParsedArgs = { dryRun: false, local: true }
-
-    await restartCluster(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-cluster')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart CloudNativePG Cluster test-cluster in namespace test-namespace',
-    )
-  })
-
-  it('should skip restart if Cluster already restarted', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    mockRestartedDeployments.add('test-namespace/test-cluster')
-
-    await restartCluster(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockD.info).not.toHaveBeenCalled()
-  })
-})
-
-describe('restartDeployment', () => {
-  let mockOwnerRef: V1OwnerReference
-  let mockRestartedDeployments: Set<string>
-  let mockD: any
-
-  beforeEach(() => {
-    mockOwnerRef = {
-      kind: 'ReplicaSet',
-      name: 'test-deployment-abcd1234',
-      apiVersion: 'apps/v1',
-      uid: 'test-uid',
-    }
-    mockRestartedDeployments = new Set<string>()
-    mockD = {
-      info: jest.fn(),
-    }
-    jest.clearAllMocks()
-  })
-
-  it('should restart Deployment when not in dry run mode', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    const mockExec = jest.fn().mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 })
-    const { $ } = require('zx')
-    ;($ as jest.Mock).mockImplementation(mockExec)
-
-    await restartDeployment(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-deployment')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith('Restarting deployment test-deployment in namespace test-namespace')
-    expect(mockD.info).toHaveBeenCalledWith('Successfully restarted deployment test-deployment')
-  })
-
-  it('should not restart Deployment in dry run mode', async () => {
-    const mockParsedArgs = { dryRun: true, local: false }
-
-    await restartDeployment(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-deployment')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart deployment test-deployment in namespace test-namespace',
-    )
-  })
-
-  it('should not restart Deployment in local mode', async () => {
-    const mockParsedArgs = { dryRun: false, local: true }
-
-    await restartDeployment(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-deployment')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart deployment test-deployment in namespace test-namespace',
-    )
-  })
-
-  it('should skip restart if Deployment already restarted', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    mockRestartedDeployments.add('test-namespace/test-deployment')
-
-    await restartDeployment(mockOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockD.info).not.toHaveBeenCalled()
-  })
-
-  it('should handle invalid ReplicaSet name gracefully', async () => {
-    const invalidOwnerRef: V1OwnerReference = {
-      kind: 'ReplicaSet',
-      name: 'invalid-name',
-      apiVersion: 'apps/v1',
-      uid: 'test-uid',
-    }
-    const mockParsedArgs = { dryRun: false, local: false }
-
-    await restartDeployment(invalidOwnerRef, 'test-namespace', mockParsedArgs, mockRestartedDeployments, mockD)
-
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Could not extract deployment name from ReplicaSet invalid-name, skipping restart',
-    )
-    expect(mockRestartedDeployments.size).toBe(0)
-  })
-})
-
-describe('restartPodOwner', () => {
-  let mockRestartedDeployments: Set<string>
-  let mockD: any
-  let mockPod: any
-
-  beforeEach(() => {
-    mockRestartedDeployments = new Set<string>()
-    mockD = {
-      info: jest.fn(),
-      warn: jest.fn(),
-    }
-    jest.clearAllMocks()
-  })
-
-  it('should restart StatefulSet owner', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    const mockExec = jest.fn().mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 })
-    const { $ } = require('zx')
-    ;($ as jest.Mock).mockImplementation(mockExec)
-
-    mockPod = {
-      metadata: {
-        namespace: 'test-namespace',
-        ownerReferences: [
-          {
-            kind: 'StatefulSet',
-            name: 'test-statefulset',
-            apiVersion: 'apps/v1',
-            uid: 'test-uid',
-          },
-        ],
-      },
-    }
-
-    await restartPodOwner(mockPod, mockD, mockParsedArgs, mockRestartedDeployments)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-statefulset')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith('Restarting StatefulSet test-statefulset in namespace test-namespace')
-  })
-
-  it('should restart CloudNativePG Cluster owner', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    const mockExec = jest.fn().mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 })
-    const { $ } = require('zx')
-    ;($ as jest.Mock).mockImplementation(mockExec)
-
-    mockPod = {
-      metadata: {
-        namespace: 'test-namespace',
-        ownerReferences: [
-          {
-            kind: 'Cluster',
-            name: 'test-cluster',
-            apiVersion: 'postgresql.cnpg.io/v1',
-            uid: 'test-uid',
-          },
-        ],
-      },
-    }
-
-    await restartPodOwner(mockPod, mockD, mockParsedArgs, mockRestartedDeployments)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-cluster')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith('Restarting CloudNativePG Cluster test-cluster in namespace test-namespace')
-  })
-
-  it('should restart Deployment via ReplicaSet owner', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    const mockExec = jest.fn().mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 })
-    const { $ } = require('zx')
-    ;($ as jest.Mock).mockImplementation(mockExec)
-
-    mockPod = {
-      metadata: {
-        namespace: 'test-namespace',
-        ownerReferences: [
-          {
-            kind: 'ReplicaSet',
-            name: 'test-deployment-abcd1234',
-            apiVersion: 'apps/v1',
-            uid: 'test-uid',
-          },
-        ],
-      },
-    }
-
-    await restartPodOwner(mockPod, mockD, mockParsedArgs, mockRestartedDeployments)
-
-    expect(mockRestartedDeployments.has('test-namespace/test-deployment')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith('Restarting deployment test-deployment in namespace test-namespace')
-  })
-
-  it('should handle pods with no owner references', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-
-    mockPod = {
-      metadata: {
-        namespace: 'test-namespace',
-      },
-    }
-
-    await restartPodOwner(mockPod, mockD, mockParsedArgs, mockRestartedDeployments)
-
-    expect(mockRestartedDeployments.size).toBe(0)
-    expect(mockD.info).not.toHaveBeenCalled()
-  })
-
-  it('should handle pods with no metadata', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-
-    mockPod = {}
-
-    await restartPodOwner(mockPod, mockD, mockParsedArgs, mockRestartedDeployments)
-
-    expect(mockRestartedDeployments.size).toBe(0)
-    expect(mockD.info).not.toHaveBeenCalled()
-  })
-
-  it('should handle multiple owner references', async () => {
-    const mockParsedArgs = { dryRun: true, local: false }
-
-    mockPod = {
-      metadata: {
-        namespace: 'test-namespace',
-        ownerReferences: [
-          {
-            kind: 'StatefulSet',
-            name: 'test-statefulset',
-            apiVersion: 'apps/v1',
-            uid: 'test-uid-1',
-          },
-          {
-            kind: 'ReplicaSet',
-            name: 'test-deployment-abcd1234',
-            apiVersion: 'apps/v1',
-            uid: 'test-uid-2',
-          },
-        ],
-      },
-    }
-
-    await restartPodOwner(mockPod, mockD, mockParsedArgs, mockRestartedDeployments)
-
-    // Should process both owner references
-    expect(mockRestartedDeployments.has('test-namespace/test-statefulset')).toBe(true)
-    expect(mockRestartedDeployments.has('test-namespace/test-deployment')).toBe(true)
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart StatefulSet test-statefulset in namespace test-namespace',
-    )
-    expect(mockD.info).toHaveBeenCalledWith(
-      'Dry run mode - would restart deployment test-deployment in namespace test-namespace',
-    )
-  })
-
-  it('should handle errors gracefully', async () => {
-    const mockParsedArgs = { dryRun: false, local: false }
-    const mockExec = jest.fn().mockRejectedValue(new Error('kubectl command failed'))
-    const { $ } = require('zx')
-    ;($ as jest.Mock).mockImplementation(mockExec)
-
-    mockPod = {
-      metadata: {
-        namespace: 'test-namespace',
-        name: 'test-pod',
-        ownerReferences: [
-          {
-            kind: 'StatefulSet',
-            name: 'test-statefulset',
-            apiVersion: 'apps/v1',
-            uid: 'test-uid',
-          },
-        ],
-      },
-    }
-
-    await restartPodOwner(mockPod, mockD, mockParsedArgs, mockRestartedDeployments)
-
-    expect(mockD.warn).toHaveBeenCalledWith(
-      'Could not restart StatefulSet for pod test-namespace/test-pod:',
-      expect.any(Error),
-    )
-  })
-})
-
-describe('detectAndRestartOutdatedIstioSidecars', () => {
-  const mockCoreApi = {
-    listPodForAllNamespaces: jest.fn(),
-  } as unknown as jest.Mocked<CoreV1Api>
-
-  const mockDeps = {
-    getDeploymentState: jest.fn(),
-    getCurrentVersion: jest.fn(),
-    loadYaml: jest.fn(),
-    getWorkloadKeyFromPod: jest.fn(),
-    restartPodOwner: jest.fn(),
-  }
-
-  beforeEach(() => {
-    jest.clearAllMocks()
-  })
-
-  it('should skip if no version upgrade', async () => {
-    mockDeps.getDeploymentState.mockResolvedValue({ version: '1.2.3' })
-    mockDeps.getCurrentVersion.mockResolvedValue('1.2.3')
-    mockCoreApi.listPodForAllNamespaces.mockResolvedValue({} as any)
-
-    await detectAndRestartOutdatedIstioSidecars(mockCoreApi, mockDeps)
-
-    expect(mockCoreApi.listPodForAllNamespaces).not.toHaveBeenCalled()
-  })
-
-  it('should log warning if expected Istio version is not found', async () => {
-    mockDeps.getDeploymentState.mockResolvedValue({})
-    mockDeps.getCurrentVersion.mockResolvedValue('2.0.0')
-    mockDeps.loadYaml.mockResolvedValue({}) // no appVersion
-    mockCoreApi.listPodForAllNamespaces.mockResolvedValue({} as any)
-
-    await detectAndRestartOutdatedIstioSidecars(mockCoreApi, mockDeps)
-
-    expect(mockCoreApi.listPodForAllNamespaces).not.toHaveBeenCalled()
-  })
-
-  it('should restart pod owners with outdated sidecars', async () => {
-    mockDeps.getDeploymentState.mockResolvedValue({ version: '1.2.3' })
-    mockDeps.getCurrentVersion.mockResolvedValue('2.0.0')
-    mockDeps.loadYaml.mockResolvedValue({ appVersion: '2.0.0' })
-    mockDeps.getWorkloadKeyFromPod.mockReturnValue('ns/deploy')
-    mockCoreApi.listPodForAllNamespaces.mockResolvedValue({
-      items: [
-        {
-          metadata: { namespace: 'ns', name: 'pod-1' },
-          spec: {
-            containers: [
-              {
-                image: 'istio/proxyv2:1.2.3',
-                name: '',
-              },
-            ],
-            initContainers: [],
-          },
-        },
-      ],
-    })
-
-    await detectAndRestartOutdatedIstioSidecars(mockCoreApi, mockDeps)
-
-    expect(mockDeps.restartPodOwner).toHaveBeenCalledTimes(1)
-  })
-
-  it('should not restart if sidecars are up to date', async () => {
-    mockDeps.getDeploymentState.mockResolvedValue({ version: '1.2.3' })
-    mockDeps.getCurrentVersion.mockResolvedValue('2.0.0')
-    mockDeps.loadYaml.mockResolvedValue({ appVersion: '2.0.0' })
-    mockDeps.getWorkloadKeyFromPod.mockReturnValue('ns/deploy')
-    mockCoreApi.listPodForAllNamespaces.mockResolvedValue({
-      items: [
-        {
-          metadata: { namespace: 'ns', name: 'pod-1' },
-          spec: {
-            containers: [
-              {
-                image: 'istio/proxyv2:2.0.0',
-                name: '',
-              },
-            ],
-            initContainers: [],
-          },
-        },
-      ],
-    })
-
-    await detectAndRestartOutdatedIstioSidecars(mockCoreApi, mockDeps)
-
-    expect(mockDeps.restartPodOwner).not.toHaveBeenCalled()
-  })
 })
 
 jest.mock('glob')
